Reward Manager: Fix - Little oopsie D:
Former-commit-id: c5d787b729aa594cb52c319857e346394e2e10a1

+++ Assets/Game/Gameplay/Rewards/RewardManager.cs
@@ -62,8 +62,8 @@ public class RewardManager : SingletonMonoBehaviour<RewardManager> {
 		get { return instance.m_pc; }
 	}
 
-	[SerializeField] private long m_xp = 0;
-	public static long xp {
+	[SerializeField] private float m_xp = 0;
+	public static float xp {
 		get { return instance.m_xp; }
 	}
